Refactor Opt Evaluator class into more smaller classes
