+++ src/main/java/de/upb/crypto/math/expressions/evaluator/MultiExpAlgorithm.java
@@ -1,0 +1,5 @@
+package de.upb.crypto.math.expressions.evaluator;
+
+public enum MultiExpAlgorithm {
+    INTERLEAVED_SLIDING, INTERLEAVED_WNAF, SIMULTANEOUS
+}
